various: compilation fixes for gcc 4.6
These changes are required to compile successfully with gcc version 4.6.

The -Wno-address option is applied to some files to bypass a warning
triggered by an rcu macro. The warning is fixed cleanly upstream, but
that fix is too risky to backport, so it is better to just silence it
this way.

Conflicts:
	arch/arm/mach-msm/Makefile